burgers/requirements.txt
asgiref==3.2.7
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.4
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.9
psycopg2==2.8.4
python-decouple==3.3
pytz==2019.3
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.8
whitenoise==5.0.1
